Doctype-class {XML}	R Documentation

Class to describe a reference to an XML DTD

Description

This class is intended to identify a DTD by SYSTEM file and/or PUBLIC catalog identifier. This is used in the DOCTYPE element of an XML document.

Objects from the Class

Objects can be created by calls to the constructor function Doctype.

Slots

name:

Object of class "character". This is the name of the top-level element in the XML document.

system:

Object of class "character". This is the name of the file on the system where the DTD document can be found. Can this be a URI?

public:

Object of class "character". This gives the PUBLIC identifier for the DTD that can be searched for in a catalog, for example to map the DTD reference to a local system element.

Methods

There is a constructor function and also methods for coerce to convert an object of this class to a character.

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.xmlsoft.org

See Also

Doctype saveXML

Examples

  d = Doctype(name = "section", public = c("-//OASIS//DTD DocBook XML V4.2//EN", "http://oasis-open.org/docbook/xml/4.2/docbookx.dtd"))  

[Package XML version 1.96-0 Index]