XMLAttributes-class {XML}	R Documentation

Class "XMLAttributes"

Description

A simple class to represent a named character vector of XML attributes some of which may have a namespace. This maintains the name space

Objects from the Class

Objects can be created by calls of the form new("XMLAttributes", ...). These are typically generated via a call to xmlAttrs.

Slots

.Data:

Object of class "character"

Extends

Class "character", from data part. Class "vector", by class "character", distance 2. Class "data.frameRowLabels", by class "character", distance 2. Class "SuperClassMethod", by class "character", distance 2.

Methods

[

signature(x = "XMLAttributes"): ...

show

signature(object = "XMLAttributes"): ...

Author(s)

Duncan Temple Lang

See Also

xmlAttrs newXMLNode xmlParse

Examples

  nn = newXMLNode("foo", attrs = c(a = "123", 'r:show' = "true"),
                   namespaceDefinitions = c(r = "http://www.r-project.org"))
  a = xmlAttrs(nn)
  a["show"]

[Package XML version 3.98-1.16 Index]