coerceNodes {XML}	R Documentation

Transform between XML representations

Description

This collection of coercion methods (i.e. as(obj, "type")) allows users of the XML package to switch between different representations of XML nodes and to map from an XML document to the root node and from a node to the document. This helps to manage the nodes

Value

An object of the target type.

See Also

xmlTreeParse xmlParse

[Package XML version 3.98-1.16 Index]