compareXMLDocs {XML}	R Documentation

Indicate differences between two XML documents

Description

This function is an attempt to provide some assistance in determining if two XML documents are the same and if not, how they differ. Rather than comparing the tree structure, this function compares the frequency distributions of the names of the node. It omits position, attributes, simple content from the comparison. Those are left to the functions that have more contextual information to compare two documents.

Usage

compareXMLDocs(a, b, ...)

Arguments

a,b
two parsed XML documents that must be internal documents, i.e. created with xmlParse or created with newXMLNode.

...
additional parameters that are passed on to the summary method for an internal document.

Value

A list with elements

inA
the names and counts of the XML elements that only appear in the first document

inB
the names and counts of the XML elements that only appear in the second document

countDiffs
a vector giving the difference in number of nodes with a particular name.

These give a description of what is missing from one document relative to the other.

Author(s)

Duncan Temple Lang

See Also

getNodeSet

Examples

tt = 
 '<x>
     <a>text</a>
     <b foo="1"/>
     <c bar="me">
        <d>a phrase</d>
     </c>
  </x>'

  a = xmlParse(tt, asText = TRUE)
  b = xmlParse(tt, asText = TRUE)
  d = getNodeSet(b, "//d")[[1]]
  xmlName(d) = "bob"
  addSibling(xmlParent(d), newXMLNode("c"))
  
  compareXMLDocs(a, b)

[Package XML version 3.98-1.16 Index]