dtdElement {XML}	R Documentation

Gets the definition of an element or entity from a DTD.

Description

A DTD in R consists of both element and entity definitions. These two functions provide simple access to individual elements of these two lists, using the name of the element or entity. The DTD is provided to determine where to look for the entry.

Usage

dtdElement(name,dtd)
dtdEntity(name,dtd)

Arguments

name 	 The name of the element being retrieved/acessed.

dtd 	 The DTD from which the element is to be retrieved.

Details

An element within a DTD contains both the list of sub-elements it can contain and a list of attributes that can be used within this tag type. dtdElement retrieves the element by name from the specified DTD definition. Entities within a DTD are like macros or text substitutes used within a DTD and/or XML documents that use it. Each consists of a name/label and a definition, the text that is substituted when the entity is referenced. dtdEntity retrieves the entity definition from the DTD. \ One can read a DTD directly (using parseDTD) or implicitly when reading a document (using xmlTreeParse) The names of all available elements can be obtained from the expression names(dtd$elements). This function is simply a convenience for indexing this elements list.

Value

An object of class XMLElementDef.

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.jclark.com/xml, http://www.omegahat.org

See Also

parseDTD, dtdValidElement

Examples

 dtdFile <- system.file("exampleData","foo.dtd", package="XML")
 foo.dtd <- parseDTD(dtdFile)
 
   # Get the definition of the `entry1' element
 tmp <- dtdElement("variable", foo.dtd)
 xmlAttrs(tmp)

 tmp <- dtdElement("entry1", foo.dtd)

  # Get the definition of the `img' entity
 dtdEntity("img", foo.dtd)

[Package XML version 1.96-0 Index]