genericSAXHandlers {XML}	R Documentation

SAX generic callback handler list

Description

This is a convenience function to get the collection of generic functions that make up the callbacks for the SAX parser. The return value can be used directly as the value of the handlers argument in xmlEventParse. One can easily specify a subset of the handlers by giving the names of the elements to include or exclude.

Usage

genericSAXHandlers(include, exclude, useDotNames = FALSE)

Arguments

include
if supplied, this gives the names of the subset of elements to return.

exclude
if supplied (and include is not), this gives the names of the elements to remove from the list of functions.

useDotNames
a logical value. If this is TRUE, the names of the elements in the list of handler functions are prefixed with '.'. This is the newer format used to differentiate general element handlers and node-name-specific handlers.

Value

A list of functions. By default, the elements are named startElement, endElement, comment, text, processingInstruction, entityDeclaration and contain the corresponding generic SAX callback function, i.e. given by the element name with the .SAX suffix.

If include or exclude is specified, a subset of this list is returned.

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.jclark.com/xml, http://www.omegahat.net

See Also

xmlEventParse startElement.SAX endElement.SAX comment.SAX processingInstruction.SAX entityDeclaration.SAX .InitSAXMethods

Examples

[Package XML version 3.98-1.16 Index]