readKeyValueDB {XML}	R Documentation

Read an XML property-list style document

Description

This function and its methods reads an XML document that is in the format of name-value or key-value pairs made up of a plist and dict nodes, each of which is made up key, and value node pairs. These used to be used for property lists on OS X and can represetn arbitrary data relatively conveniently.

Usage

readKeyValueDB(doc, ...)

Arguments

doc
the object containing the data. This can be the name of a file, a parsed XML document or an XML node.

...
additional parameters for the methods. One can pass dropComments as a logical value to control whether comment nodes are processed or ignored (TRUE).

Value

An R object representing the data read from the XML content. This is typically a named list or vector where the names are the keys and the values are collected into an R "container".

Author(s)

Duncan Temple Lang

References

Property lists.

See Also

readSolrDoc, xmlToList, xmlToDataFrame, xmlParse

Examples

 if(file.exists("/usr/share/hiutil/Stopwords.plist")) {
  o = readKeyValueDB("/usr/share/hiutil/Stopwords.plist")
 }

 if(file.exists("/usr/share/java/Tools/Applet Launcher.app/Contents/Info.plist"))
    javaInfo = readKeyValueDB('/usr/share/java/Tools/Applet Launcher.app/Contents/Info.plist')

[Package XML version 3.98-1.16 Index]