readSolrDoc {XML}	R Documentation

Read the data from a Solr document

Description

Solr documents are used to represent general data in a reasonably simple format made up of lists, integers, logicals, longs, doubles, dates, etc. each with an optional name. These correspond very naturally to R objects.

Usage

readSolrDoc(doc, ...)

Arguments

doc
the object containing the data. This can be the name of a file, a parsed XML document or an XML node.

...
additional parameters for the methods.

Value

An R object representing the data in the Solr document, typically a named vector or named list.

Author(s)

Duncan Temple Lang

References

Lucene text search system.

See Also

readKeyValueDB, xmlToList, xmlToDataFrame, xmlParse

Examples

f = system.file("exampleData", "solr.xml", package = "XML")
readSolrDoc(f)

[Package XML version 3.98-1.16 Index]