removeXMLNamespaces {XML}	R Documentation

Remove namespace definitions from a XML node or document

Description

This function and its methods allow one to remove one or more XML namespace definitions on XML nodes within a document.

Usage

removeXMLNamespaces(node, ..., all = FALSE, .els = unlist(list(...)))

Arguments

node
an XMLInternalNode or XMLInternalDocument object

...
the names of the namespaces to remove or an XMLNamespaceRef object returned via getNodeSet or xpathApply.

all
a logical value indicating whether to remove all the namespace definitions on a node.

.els
a list which is sometimes a convenient way to specify the namespaces to remove.

Value

This function is used for its side-effects and changing the internal node.

Author(s)

Duncan Temple Lang

See Also

newXMLNamespace

[Package XML version 3.98-1.16 Index]