toString.XMLNode {XML}	R Documentation

Creates string representation of XML node

Description

This creates a string from a hierarchical XML node and its children just as it prints on the console or one might see it in a document.

Usage

## S3 method for class 'XMLNode'
toString(x, ...)

Arguments

x
an object of class XMLNode.

...
currently ignored

Details

This uses a textConnection object using the name .tempXMLOutput. Since this is global, it will overwrite any existing object of that name! As a result, this function cannot be used recursively in its present form.

Value

A character vector with one element, that being the string corresponding to the XML node's contents.

Note

This requires the Expat XML parser to be installed.

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.jclark.com/xml

See Also

xmlNode xmlTreeParse

Examples

 x <- xmlRoot(xmlTreeParse(system.file("exampleData", "gnumeric.xml", package = "XML")))

 toString(x)

[Package XML version 3.98-1.16 Index]