xmlName {XML}	R Documentation

Extraces the tag name of an XMLNode object.

Description

Each XMLNode object has an element or tag name introduced in the <name ...> entry in an XML document. This function returns that name.

We can also set that name using xmlName(node) <- "name" and the value can have an XML name space prefix, e.g. "r:name".

Usage

xmlName(node, full = FALSE)

Arguments

node 	 The XMLNode object whose tag name is being requested.

full 	 a logical value indicating whether to prepend the namespace prefix, if there is one, or return just the name of the XML element/node. TRUE means prepend the prefix.

Value

A character vector of length textit{1} which is the node$name entry.

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.jclark.com/xml, http://www.omegahat.org

See Also

xmlChildren, xmlAttrs, xmlTreeParse

Examples

 fileName <- system.file("exampleData", "test.xml", package="XML") 
 doc <- xmlTreeParse(fileName)
 xmlName(xmlRoot(doc)[[1]])

 tt = xmlRoot(doc)[[1]]
 xmlName(tt)
 xmlName(tt) <- "bob"

  # We can set the node on an internal object also.
 n = newXMLNode("x")

 xmlName(n)
 xmlName(n) <- "y"

 
 xmlName(n) <- "r:y"

[Package XML version 1.96-0 Index]