xmlSearchNs {XML}	R Documentation

Find a namespace definition object by searching ancestor nodes

Description

This function allows one to search an XML tree from a particular node and find the namespace definition for a given namespace prefix or URL. This namespace definition can then be used to set it on a node to make it the effective namespace for that node.

Usage

xmlSearchNs(node, ns, asPrefix = TRUE, doc = as(node, "XMLInternalDocument"))

Arguments

node
an XMLInternaElementNode

ns
a character string (vector of length 1). If asPrefix is TRUE, this is the namespace alias/prefix. If asPrefix is FALSE, this is the URL of the namespace definition

asPrefix
a logical value. See ns.

doc
the XML document in which the node(s) are located

Value

An object of class XMLNamespaceRef.

Author(s)

Duncan Temple Lang

References

libxml2

See Also

newXMLNode

Examples

 txt = '<top xmlns:r="http://www.r-project.org"><section><bottom/></section></top>'

 doc = xmlParse(txt)
 bottom = xmlRoot(doc)[[1]][[1]]
 xmlSearchNs(bottom, "r")

[Package XML version 3.98-1.16 Index]