startElement.SAX {XML}	R Documentation

Generic Methods for SAX callbacks

Description

This is a collection of generic functions for which one can write methods so that they are called in repsonse to different SAX events. The idea is that one defines methods for different classes of the .state argument and dispatch to different methods based on that argument. The functions represent the different SAX events.

Usage

startElement.SAX(name, atts, .state = NULL)
endElement.SAX(name, .state = NULL)
comment.SAX(content, .state = NULL)
processingInstruction.SAX(target, content, .state = NULL)
text.SAX(content, .state = NULL)
entityDeclaration.SAX(name, base, sysId, publicId, notationName,  .state = NULL)
.InitSAXMethods(where = "package:XML")

Arguments

name 	 the name of the XML element or entity being declared

atts

content 	 the value/string in the processing instruction or comment

target 	 the target of the processing instruction, e.g. the R in <?R....>

base 	 x

sysId 	 the system identifier for this entity

publicId 	 the public identifier for the entity

notationName 	 name of the notation specification

.state 	 the state object on which the user-defined methods should dispatch.

where 	 the package in which the class and method definitions should be defined. This is almost always unspecified.

Value

Each method should return the (potentially modified) state value.

Note

This no longer requires the Expat XML parser to be installed. Instead, we use libxml's SAX parser.

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.xmlsoft.org

See Also

xmlEventParse

[Package XML version 1.96-0 Index]