XMLNode-class {XML}	R Documentation

Classes to describe an XML node object.

Description

These classes are intended to represent an XML node, either directly in S or a reference to an internal libxml node. Such nodes respond to queries about their name, attributes, namespaces and children. These are old-style, S3 class definitions at present.

Slots

Methods

No methods defined with class "XMLNode" in the signature.

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.xmlsoft.org

See Also

xmlTreeParse xmlTree newXMLNode xmlNode

Examples

 # An R-level XMLNode object
a <- xmlNode("arg", attrs = c(default="T"),
               xmlNode("name", "foo"), xmlNode("defaultValue","1:10"))
  

[Package XML version 1.96-0 Index]