docName {XML}	R Documentation

Accessors for name of XML document

Description

These functions and methods allow us to query and set the “name” of an XML document. This is intended to be its URL or file name or a description of its origin if raw XML content provided as a string.

Usage

docName(doc)

Arguments

doc 	 the XML document object, of class XMLInternalDocument or XMLDocument.

Value

A character string giving the name. If the document was created from text, this is NA (of class character).
The assignment function returns the updated object, but the R assignment operation will return the value on the right of the assignment!

Author(s)

Duncan Temple Lang

See Also

xmlTreeParse xmlInternalTreeParse newXMLDoc

Examples

  f = system.file("exampleData", "catalog.xml",  package = "XML")
  doc = xmlInternalTreeParse(f)
  docName(doc)

  doc = xmlInternalTreeParse("<a><b/></a>", asText = TRUE)
      # an NA
  docName(doc)
  docName(doc) = "Simple XML example"
  docName(doc)

[Package XML version 1.96-0 Index]