libxmlVersion {XML}	R Documentation

Query the version and available features of the libxml library.

Description

libxmlVersion retrieves the version of the libxml library used when installing this XML package. libxmlFeatures returns a named logical vector indicating which features are enabled.

Usage

libxmlVersion(runTime = FALSE)
libxmlFeatures()

Arguments

runTime
a logical value indicating whether to retrieve the version information describing libxml when the R package was compiled or the run-time version. These may be different if a) a new version of libxml2 is installed after the package is installed, b) if the package was installed as a binary package built on a different machine.

Value

libxmlVersion returns a named list with fields

major
the major version number, either 1 or 2 indicating the old or new-style library.

minor
the within version release number.

patch
the within minor release version number

libxmlFeatures returns a logical vector with names given by: [1] "THREAD" "TREE" "OUTPUT" "PUSH" "READER" [6] "PATTERN" "WRITER" "SAX1" "FTP" "HTTP" [11] "VALID" "HTML" "LEGACY" "C14N" "CATALOG" [16] "XPATH" "XPTR" "XINCLUDE" "ICONV" "ISO8859X" [21] "UNICODE" "REGEXP" "AUTOMATA" "EXPR" "SCHEMAS" [26] "SCHEMATRON" "MODULES" "DEBUG" "DEBUG_MEM" "DEBUG_RUN" [31] "ZLIB" Elements are either TRUE or FALSE indicating whether support was activatd for that feature, or NA if that feature is not part of the particular version of libcurl.

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.xmlsoft.org, http://www.omegahat.net

Examples

 ver <- libxmlVersion()
 if(is.null(ver)) {
   cat("Relly old version of libxml\n")
 } else {
   if(ver$major > 1) {
     cat("Using libxml2\n")
   }
 }

[Package XML version 3.98-1.16 Index]