libxmlVersion {XML}	R Documentation

Get the version of the libxml library.

Description

This currently retrieves the version of the libxml library used when installing this XML package.

Usage

libxmlVersion()

Details

This computes the version at compile time. In the future we will hopefully find a way to find the run-time version.

Value

A named list with fields

major 	 the major version number, either 1 or 2 indicating the old or new-style library.

minor 	 the within version release number.

patch 	 the within minor release version number

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.xmlsoft.org, http://www.omegahat.org

Examples

 ver <- libxmlVersion()
 if(is.null(ver)) {
   cat("Relly old version of libxml\n")
 } else {
   if(ver$major > 1) {
     cat("Using libxml2\n")
   }
 }

[Package XML version 1.96-0 Index]