parseURI {XML}	R Documentation

Parse a URI string into its elements

Description

This breaks a URI given as a string into its different elements such as protocol/scheme, host, port, file name, query. This information can be used, for example, when constructing URIs relative to a base URI.

The return value is an S3-style object of class URI.

This function uses libxml routines to perform the parsing.

Usage

parseURI(uri)

Arguments

uri 	 a single string

Value

A list with 8 elements

scheme 	 the name of the protocol being used, http, ftp as a string.

authority 	 a string represeting a rarely used aspect of URIs

server 	 a string identifying the host, e.g. www.omegahat.org

user 	 a string giving the name of the user, e.g. in FTP "ftp://duncan@www.omegahat.org", this would yield "duncan"

path 	 a string identifying the path of the target file

query 	 the CGI query part of the string, e.g. the bit after '?' of the form name=value&name=value

fragment 	 a string giving the coo

port 	 an integer identifying the port number on which the connection is to be made

Examples

  parseURI("http://www.omegahat.org:8080/RCurl/index.html")
  parseURI("ftp://duncan@www.omegahat.org:8080/RCurl/index.html") 

[Package XML version 1.96-0 Index]