xmlAttributeType {XML}	R Documentation

The type of an XML attribute for element from the DTD

Description

This examines the definition of the attribute, usually returned by parsing the DTD with parseDTD and determines its type from the possible values: Fixed, string data, implied, required, an identifier, an identifier reference, a list of identifier references, an entity, a list of entities, a name, a list of names, an element of enumerated set, a notation entity.

Usage

xmlAttributeType(def, defaultType=FALSE)

Arguments

def 	 the attribute definition object, usually retrieved from the DTD via parseDTD.

defaultType 	 whether to return the default value if this attribute is defined as being a value from an enumerated set.

Value

A string identifying the type for the sspecified attributed.

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.omegahat.org/RSXML

See Also

parseDTD

[Package XML version 1.96-0 Index]