xmlEventHandler {XML}	R Documentation

Default handlers for the SAX-style event XML parser

Description

This is a function that returns a closure instance containing the default handlers for use with xmlEventParse for parsing XML documents via the SAX-style parsing.

Usage

xmlEventHandler()

Details

These handlers simply build up the DOM tree and thus perform the same job as xmlTreeParse. It is here more as an example, reference and a base that users can extend.

Value

The return value is a list of functions which are used as callbacks by the internal XML parser when it encounters certain XML elements/structures. These include items such as the start of an element, end of an element, processing instruction, text node, comment, entity references and definitions, etc.

startElement

endElement

processingInstruction

text

comment

externalEntity

entityDeclaration

cdata

dom

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.jclark.com/xml, http://www.omegahat.org

See Also

xmlEventParse xmlTreeParse

Examples

 xmlEventParse(system.file("exampleData", "mtcars.xml", package="XML"),
               handlers=xmlEventHandler())

[Package XML version 1.96-0 Index]