xmlNamespace {XML}	R Documentation

Retrieve the namespace value of an XML node.

Description

Each XML node has a namespace identifier which is a string indicating in which DTD (Document Type Definition) the definition of that element can be found. This avoids the problem of having different document definitions using the same names for XML elements that have different meaning. To resolve the name space, i.e. i.e. find out to where the identifier points, one can use the expression xmlNamespace(xmlRoot(doc)).

The class of the result is is an S3-style object of class XMLNamespace.

Usage

xmlNamespace(x)

Arguments

x 	 the object whose namespace is to be computed

Value

For non-root nodes, this returns a string giving the identifier of the name space for this node. For the root node, this returns a list with 2 elements:

id 	 the identifier by which other nodes refer to this namespace.

uri 	 the URI or location that defines this namespace.

local 	 ? (can't remember off-hand).

Author(s)

Duncan Temple Lang

References

http://www.w3.org/XML, http://www.jclark.com/xml, http://www.omegahat.org

See Also

xmlName xmlChildren xmlAttrs xmlValue xmlNamespaceDefinitions

Examples

  doc <- xmlTreeParse(system.file("exampleData", "job.xml", package="XML"))
  xmlNamespace(xmlRoot(doc))
  xmlNamespace(xmlRoot(doc)[[1]][[1]])

  doc <- xmlInternalTreeParse(system.file("exampleData", "job.xml", package="XML"))
     # Since the first node is a comment, get the sibling.
  xmlNamespace(getSibling(xmlRoot(doc)))
  xmlNamespace(getSibling(xmlRoot(doc))[[1]][[1]])

  node <- xmlNode("arg", xmlNode("name", "foo"), namespace="R")
  xmlNamespace(node)

[Package XML version 1.96-0 Index]